requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
Django==5.0.4
django-advanced-password-validation==1.2.0
django-anymail==12.0
django-password-strength==1.2.1
django-phone-field==1.8.1
django-social-share==2.3.0
django-taggit==6.1.0
django-tinymce==4.0.0
django-use-email-as-username==1.4.0
idna==3.10
pillow==10.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyotp==2.9.0
python-decouple==3.8
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.3
